added more support for oft node interface and set up structures uio and iovec to copy data to and from user to kernel

+++ kern/include/fdtable.h
@@ -24,5 +24,5 @@ fdtable* fdtable_init(void);
 void fdtable_destroy(fdtable *fdt);
 int stdio_init(fdtable* fdt);
 int do_sys_open(const_userptr_t path, int flags, mode_t mode, int* retval);
-int do_sys_write(int fd, const_userptr_t buf, size_t nbytes, int *retval);
+ssize_t do_sys_write(int fd, const_userptr_t buf, size_t nbytes, int *retval);
 #endif
